Add dashboard component
